Refactor interact button and move to separate file
